Tidied code, made code style more consistent with older libspatialaudio style
